requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
fluent.runtime==0.3.1
fluent.syntax==0.18.1
frozenlist==1.4.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==1.10.5
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.9.0
yarl==1.9.4
